requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
psycopg2-binary==2.9.9
python-dotenv == 1.0.1